+++ src/iLogger.js
@@ -1,3 +1,0 @@
-//
-// Exporting an array of method names that a logger should implement.
-export const iLogger = ["log", "info", "warn", "error"];